refactor: major to split up stages
