+++ oceanarray/analysis/__init__.py
@@ -1,0 +1,7 @@
+"""Analysis subpackage for oceanarray.
+
+Submodules
+----------
+_science : Science utility functions (lag_correlation, isopycnal tracking,
+           wavelet analysis, etc.) — promoted from the top-level tools.py.
+"""
