+++ oceanarray/config/__init__.py
@@ -1,0 +1,7 @@
+"""Configuration and constants for oceanarray.
+
+Submodules
+----------
+parameters : Package-level constants, QC thresholds, and colormap defaults.
+validation : Mooring YAML validation utilities.
+"""
