+++ oceanarray/mooring/__init__.py
@@ -1,0 +1,1 @@
+"""Mooring-level processing: stacking and gridding multiple instruments."""
