+++ oceanarray/mooring/grid.py
@@ -1,32 +1,284 @@
-"""Step 1 processing for mooring data: Time gridding and optional filtering.
-
-This module handles:
-- Loading processed Stage 2 NetCDF files (_stage2.nc) from multiple instruments
-- Optional time-domain filtering applied to individual instrument records
-- Interpolating all instruments onto a common time grid
-- Combining instruments into a single dataset with N_LEVELS dimension
-- Encoding instrument metadata as coordinate arrays
-- Writing time-gridded mooring datasets
-
-This represents Step 1 in the mooring-level processing workflow:
-- Step 1: Time gridding (this module)
-- Step 2: Vertical gridding (future)
-- Step 3: Multi-deployment stitching (future)
-
-IMPORTANT: Filtering is applied to individual instrument records BEFORE interpolation
-to preserve data integrity and avoid interpolation artifacts.
-
-Version: 1.1
-Last updated: 2025-09-07
-"""
+"""MooringGridder and TimeGriddingProcessor: grid mooring data onto regular time/depth axes."""
 
 from pathlib import Path
 from typing import Any, Dict, List, Optional, Tuple
-
 import numpy as np
 import xarray as xr
 import yaml
 from seasenselib.writers import NetCdfWriter
+from oceanarray.utilities import _status, cast_output_dtypes
+from oceanarray.mooring.helpers import _get_proc_dir
+
+
+class MooringGridder:
+    """Step 2: vertically interpolate stacked instruments onto a pressure grid → ``_grid.nc``."""
+
+    def __init__(
+        self,
+        base_dir: Optional[str] = None,
+        *,
+        proc_dir: Optional[str] = None,
+    ) -> None:
+        """Interpolate a mooring stack onto a regular pressure grid.
+
+        Reads ``{mooring}_stack.nc`` (produced by :class:`MooringStacker`) and
+        interpolates all scalar variables onto a uniform pressure axis defined by
+        ``p_start``, ``p_end``, and ``dp`` (in dbar).  Writes
+        ``{mooring}_grid.nc`` in the mooring proc directory.
+
+        Parameters
+        ----------
+        base_dir : str, optional
+            Legacy: cruise-level base directory containing a ``proc/`` subdirectory.
+        proc_dir : str, optional
+            Cruise-level processed output directory. Pipeline appends ``/{mooring}/``.
+
+        """
+        if base_dir is not None:
+            self.base_dir: Optional[Path] = Path(base_dir)
+            self._proc_dir: Optional[Path] = None
+            self._legacy = True
+        else:
+            self.base_dir = None
+            self._proc_dir = Path(proc_dir) if proc_dir else None
+            self._legacy = False
+
+    def _resolve_proc_dir(self, mooring_name: str) -> Path:
+        """Return the mooring-level proc directory."""
+        if not self._legacy and self._proc_dir is not None:
+            return self._proc_dir / mooring_name
+        return _get_proc_dir(self.base_dir, mooring_name)
+
+    def _rel(self, path: Path) -> str:
+        """Return a short display path relative to base_dir or proc_dir."""
+        for root in (r for r in (self.base_dir, self._proc_dir) if r):
+            try:
+                return str(path.relative_to(root))
+            except ValueError:
+                continue
+        return path.name
+
+    def grid(
+        self,
+        mooring_name: str,
+        p_start: float = 200.0,
+        p_end: float = 1000.0,
+        dp: float = 20.0,
+        force: bool = False,
+    ) -> bool:
+        """Interpolate the stacked mooring dataset onto a regular pressure grid.
+
+        Reads ``{mooring}_stack.nc``, interpolates all variables onto a pressure
+        axis from *p_start* to *p_end* in steps of *dp* dbar, and writes
+        ``{mooring}_grid.nc``.
+
+        Returns True on success, False on error.
+        """
+        proc_dir = self._resolve_proc_dir(mooring_name)
+        merge_path = proc_dir / f"{mooring_name}_stack.nc"
+        output_path = proc_dir / f"{mooring_name}_grid.nc"
+
+        try:
+            stack_found = merge_path.exists()
+            output_found = output_path.exists()
+        except (TimeoutError, OSError) as exc:
+            print("ERROR: Cannot access data drive — is it connected?")
+            print(f"       Path: {merge_path.parent}")
+            print(f"       ({type(exc).__name__}: {exc})")
+            return False
+
+        if not stack_found:
+            print(f"ERROR: Stack file not found: {merge_path}")
+            print("       Run 'oceanarray stack' first.")
+            return False
+
+        if output_found and not force:
+            _status("skip", self._rel(output_path))
+            return True
+
+        p_grid = np.arange(p_start, p_end + dp * 0.5, dp)
+        n_p = len(p_grid)
+
+        try:
+            ds = xr.open_dataset(merge_path).load()
+        except (TimeoutError, OSError) as exc:
+            print("ERROR: Cannot read stack file — is the data drive connected?")
+            print(f"       ({type(exc).__name__}: {exc})")
+            return False
+        n_time = ds.sizes["time"]
+        n_instr = ds.sizes["N_LEVELS"]
+
+        print(
+            f"Gridding {n_instr} instruments → {n_p} pressure levels "
+            f"({p_start:.0f}:{dp:.0f}:{p_end:.0f} dbar) × {n_time} time steps"
+        )
+
+        if "pressure" not in ds.data_vars:
+            print("ERROR: 'pressure' not found in stack file — cannot grid vertically")
+            ds.close()
+            return False
+
+        pressure = ds["pressure"].values.astype(np.float64)  # (time, N_LEVELS)
+        # Variables that are meaningful at the stacked per-instrument level but
+        # should not be interpolated onto the pressure grid.  Instrument-frame
+        # and beam-frame velocities are excluded because the XYZ→ENU rotation
+        # has already been applied; gridding the pre-rotation components would
+        # be misleading.  Diagnostic quantities (amplitude, correlation,
+        # battery) are per-sensor and do not have a physical meaning on a
+        # spatially-interpolated grid.
+        _GRID_EXCLUDE: frozenset = frozenset(
+            {
+                "velocity_x",
+                "velocity_y",
+                "velocity_z",
+                "velocity_beam1",
+                "velocity_beam2",
+                "velocity_beam3",
+                "amplitude",
+                "correlation",
+                "battery_voltage",
+                "velocity_flag",  # flag array, not a gridded physics variable
+                "tilt_from_pressure",  # per-instrument diagnostic, not gridded
+                "tilt_pressure_ref_hab",
+                "heading",  # instrument-frame orientation — not meaningful on a pressure grid
+                "pitch",
+                "roll",
+            }
+        )
+        grid_vars = [
+            v
+            for v in ds.data_vars
+            if v != "pressure"
+            and ds[v].dims == ("time", "N_LEVELS")
+            and not v.endswith("_qc")
+            and v not in _GRID_EXCLUDE
+        ]
+
+        stacked: Dict[str, np.ndarray] = {
+            v: np.full((n_p, n_time), np.nan) for v in grid_vars
+        }
+        var_data = {v: ds[v].values.astype(np.float64) for v in grid_vars}
+
+        # QC masking before vertical interpolation.
+        # T/S/P: NaN any non-finite values (already masked at stack time).
+        # Velocity: use velocity_flag from the stack to NaN suspect/bad samples.
+        #   Flag 3 (suspect), 4 (bad), 9 (missing) → NaN before interpolation.
+        _GRIDDER_TSQC = {"temperature", "conductivity", "salinity"}
+        for _v in grid_vars:
+            if _v in _GRIDDER_TSQC:
+                var_data[_v][~np.isfinite(var_data[_v])] = np.nan
+
+        _vel_vars = {
+            "east_velocity",
+            "north_velocity",
+            "up_velocity",
+            "current_speed",
+            "current_direction",
+        }
+        if "velocity_flag" in ds.data_vars:
+            _vflag = ds["velocity_flag"].values.astype(np.float64)
+            _bad_vel = np.isin(np.round(_vflag).astype(np.int8), [3, 4, 9])
+            for _v in _vel_vars:
+                if _v in var_data:
+                    var_data[_v][_bad_vel] = np.nan
+        else:
+            for _v in _vel_vars:
+                if _v in var_data:
+                    var_data[_v][~np.isfinite(var_data[_v])] = np.nan
+
+        # ADCP standalone QC variables — mask velocity before gridding.
+        # seabed_qc removes bins at/below the seafloor; percent_good_qc and
+        # error_velocity_qc remove poor-quality pings.  All three use flag >= 3
+        # (suspect or worse) as the NaN threshold, consistent with how the
+        # rose diagram and stack report mask these variables.
+        for _adcp_qc in ("seabed_qc", "percent_good_qc", "error_velocity_qc"):
+            if _adcp_qc not in ds.data_vars:
+                continue
+            _qc = ds[_adcp_qc].values  # (time, N_LEVELS) after stack transpose
+            _bad = np.isin(np.round(_qc).astype(np.int8), [3, 4, 9])
+            for _v in _vel_vars:
+                if _v in var_data:
+                    var_data[_v][_bad] = np.nan
+
+        for t in range(n_time):
+            p_col = pressure[t, :]
+            p_valid_mask = np.isfinite(p_col)
+            if p_valid_mask.sum() < 2:
+                continue
+
+            for vname in grid_vars:
+                v_col = var_data[vname][t, :]
+                both_valid = p_valid_mask & np.isfinite(v_col)
+                if both_valid.sum() < 2:
+                    continue
+                p_v = p_col[both_valid]
+                v_v = v_col[both_valid]
+                sort_idx = np.argsort(p_v)
+                stacked[vname][:, t] = np.interp(
+                    p_grid,
+                    p_v[sort_idx],
+                    v_v[sort_idx],
+                    left=np.nan,
+                    right=np.nan,
+                )
+
+        # Build output dataset; skip variables that are entirely NaN
+        data_vars: Dict = {}
+        for vname in grid_vars:
+            if not np.all(np.isnan(stacked[vname])):
+                a = dict(ds[vname].attrs)
+                a["vertical_interpolation"] = (
+                    f"linear in pressure; no extrapolation outside {p_start:.0f}–{p_end:.0f} dbar"
+                )
+                data_vars[vname] = xr.Variable(
+                    ("time", "pressure"), stacked[vname].T, attrs=a
+                )
+
+        ds_out = xr.Dataset(
+            data_vars,
+            coords={
+                "time": ds["time"],
+                "pressure": xr.Variable(
+                    "pressure",
+                    p_grid,
+                    {
+                        "units": "dbar",
+                        "long_name": "sea water pressure",
+                        "standard_name": "sea_water_pressure",
+                        "axis": "Z",
+                        "positive": "down",
+                    },
+                ),
+            },
+        )
+
+        prior_history = ds.attrs.get("history", "")
+        ds_out.attrs.update(
+            {
+                **{k: v for k, v in ds.attrs.items() if k not in ("history",)},
+                "p_start_dbar": p_start,
+                "p_end_dbar": p_end,
+                "dp_dbar": dp,
+                "history": (
+                    prior_history
+                    + f"; Step 2 grid: linear interpolation in pressure onto {dp:.0f} dbar pressure grid "
+                    f"({p_start:.0f}–{p_end:.0f} dbar); no extrapolation"
+                ),
+            }
+        )
+
+        ds.close()
+        if output_path.exists():
+            output_path.unlink()
+        ds_out = cast_output_dtypes(ds_out)
+        _enc = {
+            v: {"zlib": True, "complevel": 5}
+            for v in ds_out.data_vars
+            if ds_out[v].dtype.kind not in ("O", "U", "S")
+        }
+        ds_out.to_netcdf(output_path, encoding=_enc)
+        _status("file", self._rel(output_path))
+        return True
 
 
 class TimeGriddingProcessor:
@@ -39,7 +291,7 @@ class TimeGriddingProcessor:
 
     def _setup_logging(self, mooring_name: str, output_path: Path) -> None:
         """Set up logging for the processing run using global config."""
-        from .logger import setup_stage_logging
+        from oceanarray.logger import setup_stage_logging
 
         self.log_file = setup_stage_logging(mooring_name, "time_gridding", output_path)
 
@@ -930,29 +1182,3 @@ def process_multiple_moorings_time_gridding(
         )
 
     return results
-
-
-# Example usage
-if __name__ == "__main__":
-    # Your mooring list
-    moorlist = ["dsE_1_2018"]
-
-    basedir = "/Users/eddifying/Dropbox/data/ifmro_mixsed/ds_data_eleanor/"
-
-    # Process all moorings without filtering
-    results = process_multiple_moorings_time_gridding(moorlist, basedir)
-
-    # Example: Process with low-pass filtering (RAPID-style de-tiding)
-    # results = process_multiple_moorings_time_gridding(
-    #     moorlist, basedir,
-    #     filter_type='lowpass',
-    #     filter_params={'cutoff_days': 2.0, 'order': 6}
-    # )
-
-    # Print summary
-    print(f"\n{'=' * 50}")
-    print("STEP 1 (TIME GRIDDING) PROCESSING SUMMARY")
-    print(f"{'=' * 50}")
-    for mooring, success in results.items():
-        status = "SUCCESS" if success else "FAILED"
-        print(f"{mooring}: {status}")
