+++ oceanarray/mooring/stack.py
@@ -1,158 +1,32 @@
-"""Mooring-level processing: instrument stacking and vertical gridding.
-
-Step 1 — stack (``oceanarray stack``):
-    Read all instruments from a single deployment and stack them onto a common
-    60 s (or user-specified) time axis.  Output: ``{mooring}_stack.nc`` with
-    dimensions ``(instrument, time)``. --> change to (time, instrument)!
-
-Step 2 — grid (``oceanarray grid``):
-    Vertically interpolate the stacked dataset onto a standard pressure grid.
-    Output: ``{mooring}_grid.nc`` with dimensions ``(pressure, time)``.
-"""
+"""MooringStacker: interpolate all instruments on a mooring onto a common time grid."""
 
 from pathlib import Path
-from typing import Any, Dict, List, Optional
-import re
-
+from typing import Dict, List, Optional
 import numpy as np
 import xarray as xr
 import yaml
-from . import parameters as P
-from .utilities import _status, cast_output_dtypes, drop_all_zero_vars
-
-# Canonical allowed values for the ``instrument:`` field in mooring YAML files.
-# Defined in parameters.py alongside the instrument→file_type mapping.
-KNOWN_INSTRUMENT_TYPES: frozenset = P.KNOWN_INSTRUMENT_TYPES
-
-STACK_VARS = [
-    "temperature",
-    "temperature_qc",
-    "salinity",
-    "salinity_qc",
-    "conductivity",
-    "pressure",
-    "pressure_qc",
-    "east_velocity",
-    "north_velocity",
-    "up_velocity",
-    "current_speed",
-    "current_direction",
-    "east_velocity_qc",
-    "north_velocity_qc",
-    "up_velocity_qc",
-    "pitch_qc",
-    "roll_qc",
-    "velocity_beam1",
-    "velocity_beam2",
-    "velocity_beam3",
-    "heading",
-    "pitch",
-    "roll",
-    "analog_input_1",
-    "analog_input_2",
-    "percent_good_qc",
-    "error_velocity_qc",
-    "seabed_qc",
-    "turbidity",
-    "turbidity_qc",
-    "dissolved_oxygen",
-    "dissolved_oxygen_qc",
-    "oxygen_saturation_pct",
-    "apparent_oxygen_utilization",
-]
-
-# Variables passed through without QC masking at the stack step.
-# Velocity and orientation are kept at full precision so downstream users
-# (grid step, analysis) can apply their own masking via velocity_flag.
-# QC flag arrays never need masking (there is no companion *_qc_qc* variable).
-_STACK_RAW: frozenset = frozenset(
-    {
-        "east_velocity",
-        "north_velocity",
-        "up_velocity",
-        "current_speed",
-        "current_direction",
-        "velocity_beam1",
-        "velocity_beam2",
-        "velocity_beam3",
-        "heading",
-        "pitch",
-        "roll",
-        "east_velocity_qc",
-        "north_velocity_qc",
-        "up_velocity_qc",
-        "pitch_qc",
-        "roll_qc",
-        "percent_good_qc",
-        "error_velocity_qc",
-        "seabed_qc",
-    }
+from oceanarray import parameters as P
+from oceanarray.utilities import (
+    _status,
+    cast_output_dtypes,
+    drop_all_zero_vars,
+    extract_inline_instruments,
+    parse_latlon,
+)
+from oceanarray.mooring.helpers import (
+    STACK_VARS,
+    _safe_serial,
+    _worst_flag,
+    _get_proc_dir,
+    _best_nc,
+    _detect_interval_s,
+    _make_adcp_head_ds,
+    _make_adcp_bin_ds,
+    _nearest_subsample,
+    _linear_interp,
 )
 
-
-def _safe_serial(serial: Any) -> str:
-    """Sanitise a serial number string for use in filenames.
-
-    If the raw YAML value contains a comma (e.g. ``16430, R01-024``), only
-    the first comma-separated token is used — the remainder is a beacon ID
-    or annotation and is not part of the filename.
-    """
-    s = str(serial).split(",")[0]
-    return re.sub(r"[^\w\-]", "", s)
-
-
-def _dms_to_deg(s: str) -> float:
-    """Parse 'DD MM.mmm N/S/E/W' or a plain float string to decimal degrees."""
-    s = s.strip()
-    try:
-        return float(s)
-    except ValueError:
-        pass
-    parts = s.upper().split()
-    hemi = parts[-1] if parts[-1] in ("N", "S", "E", "W") else None
-    nums = parts[:-1] if hemi else parts
-    val = sum(float(x) / 60.0**i for i, x in enumerate(nums))
-    if hemi in ("S", "W"):
-        val = -val
-    return val
-
-
-def _parse_latlon(cfg: dict) -> tuple[float, float]:
-    """Return (lat, lon) in decimal degrees from a mooring config dict."""
-    lat, lon, _ = _parse_latlon_with_source(cfg)
-    return lat, lon
-
-
-def _parse_latlon_with_source(cfg: dict) -> tuple[float, float, str]:
-    """Return (lat, lon, source_key) from a mooring config dict.
-
-    source_key is the YAML key pair used, e.g. 'seabed_latitude/seabed_longitude',
-    or 'unknown (defaulting to 0, 0)' if none found.
-
-    Zero values (lat == 0 and lon == 0) are treated as unfilled placeholders and
-    skipped so that a later key with a real location is used instead.  This handles
-    YAML templates where fields like ``seabed_latitude: 00 00.000 N`` are present
-    but not yet populated.
-    """
-    for lat_key, lon_key in [
-        ("seabed_latitude", "seabed_longitude"),
-        ("deployment_latitude", "deployment_longitude"),
-        ("planned_latitude", "planned_longitude"),
-        ("latitude", "longitude"),
-    ]:
-        lat_s = cfg.get(lat_key)
-        lon_s = cfg.get(lon_key)
-        if lat_s is not None and lon_s is not None:
-            try:
-                lat_v = _dms_to_deg(str(lat_s))
-                lon_v = _dms_to_deg(str(lon_s))
-            except Exception:  # noqa: BLE001  — _dms_to_deg parse failure; try next key
-                continue
-            if lat_v == 0.0 and lon_v == 0.0:
-                continue  # treat (0, 0) as an unfilled placeholder
-            return lat_v, lon_v, f"{lat_key}/{lon_key}"
-    return 0.0, 0.0, "unknown (defaulting to 0, 0)"
-
+KNOWN_INSTRUMENT_TYPES: frozenset = P.KNOWN_INSTRUMENT_TYPES
 
 _SIGMA_META = {
     0: (
@@ -181,289 +55,6 @@ _SIGMA_META = {
         "sea_water_sigma_4",
     ),
 }
-
-
-def _apply_qc_mask(src_v: np.ndarray, ds: "xr.Dataset", vname: str) -> np.ndarray:
-    """Return src_v with flagged values replaced by NaN.
-
-    Flags kept as valid:
-      0 (no QC performed), 1 (good), 2 (probably good)
-      + flag 8 (interpolated) for pressure only.
-    Flags masked (→ NaN):
-      3 (suspect), 4 (bad), 9 (missing value).
-    If no companion *_qc* variable exists the array is returned unchanged.
-    """
-    qc_name = f"{vname}_qc"
-    if qc_name not in ds.data_vars:
-        return src_v
-    qc = ds[qc_name].values
-    if vname == "pressure":
-        keep = np.isin(qc, [0, 1, 2, 8])
-    else:
-        keep = np.isin(qc, [0, 1, 2])
-    out = src_v.copy()
-    out[~keep] = np.nan
-    return out
-
-
-def _worst_flag(a: np.ndarray, b: np.ndarray) -> np.ndarray:
-    """Element-wise worst QC flag (9 > 4 > 3 > 8 > 2 > 1 > 0).
-
-    NaN inputs are treated as flag 9 (missing value) so that levels with no
-    data are never silently promoted to flag 0 ("no QC performed").
-    """
-    # rank[flag_value] gives priority; higher rank = worse flag
-    _rank = np.array([0, 1, 2, 4, 5, 6, 6, 6, 3, 7], dtype=np.int8)
-    a = np.where(np.isfinite(a), a, 9.0)
-    b = np.where(np.isfinite(b), b, 9.0)
-    ai = np.clip(np.round(a).astype(np.int8), 0, 9)
-    bi = np.clip(np.round(b).astype(np.int8), 0, 9)
-    take_b = _rank[bi] > _rank[ai]
-    out = ai.copy()
-    out[take_b] = bi[take_b]
-    return out
-
-
-def _times_to_float(t: np.ndarray) -> np.ndarray:
-    return t.astype("datetime64[ns]").astype(np.float64)
-
-
-def _get_proc_dir(base_dir: Path, mooring_name: str) -> Path:
-    proc = base_dir / "proc"
-    if not proc.is_dir():
-        legacy = base_dir / "moor" / "proc"
-        proc = legacy if legacy.is_dir() else proc
-    return proc / mooring_name
-
-
-def _best_nc(
-    proc_dir: Path, instr_type: str, mooring_name: str, serial: str
-) -> Optional[Path]:
-    """Return best available stage3 or stage2 file for one instrument; None if only stage1."""
-    base = proc_dir / instr_type / f"{mooring_name}_{serial}"
-    for suffix in ("_stage3.nc", "_stage2.nc"):
-        p = Path(str(base) + suffix)
-        if p.exists():
-            return p
-    return None
-
-
-def _detect_interval_s(time_vals: np.ndarray) -> float:
-    if len(time_vals) < 2:
-        return 60.0
-    dt = np.diff(time_vals.astype("datetime64[s]").astype(np.float64))
-    return float(np.median(dt))
-
-
-# Variables measured at the ADCP transducer head (point measurements, not per-bin).
-# These must NOT appear in per-bin datasets — they belong only in the head entry.
-_ADCP_HEAD_VARS: frozenset = frozenset(
-    {
-        "temperature",
-        "temperature_qc",
-        "heading",
-        "pitch",
-        "roll",
-        "battery_voltage",
-        "speed_of_sound",
-        "analog_input_1",
-        "analog_input_2",
-    }
-)
-
-
-def _make_adcp_head_ds(ds_parent: xr.Dataset) -> xr.Dataset:
-    """Create a 1-D (time-only) view of the ADCP transducer head for stack insertion.
-
-    The ADCP transducer head is treated as a distinct instrument position in the mooring
-    stack.  It carries instrument-diagnostic variables that describe the sensor, not the
-    water column:
-
-    - ``temperature`` (°C) — internal electronics temperature; useful for diagnosing
-      warm-up drift but not a reliable water temperature measurement.
-    - ``heading``, ``pitch``, ``roll`` (°) — instrument orientation.
-    - ``pressure`` (dbar) — pressure at the transducer face, i.e. the instrument depth.
-
-    Per-bin velocity and ``bin_pressure`` are deliberately excluded.  Each velocity bin
-    gets its own stack entry at the pressure appropriate to that bin
-    (see ``_make_adcp_bin_ds``), rather than the transducer pressure.
-
-    Parameters
-    ----------
-    ds_parent : xr.Dataset
-        Full ADCP stage-3 dataset, including 2-D ``(time, N_BINS)`` variables.
-
-    Returns
-    -------
-    xr.Dataset
-        Dataset containing only time-series (1-D) variables from ``_ADCP_HEAD_VARS``
-        plus the transducer ``pressure``.  Suitable for merging into a point-instrument
-        stack.
-
-    """
-    keep = [
-        v
-        for v in ds_parent.data_vars
-        if v in _ADCP_HEAD_VARS and ds_parent[v].dims == ("time",)
-    ]
-    # Also keep the instrument-head pressure if present (1-D, not bin_pressure)
-    if "pressure" in ds_parent.data_vars and ds_parent["pressure"].dims == ("time",):
-        if "pressure" not in keep:
-            keep.append("pressure")
-    return ds_parent[keep]
-
-
-def _make_adcp_bin_ds(ds_parent: xr.Dataset, bin_idx: int) -> xr.Dataset:
-    """Create a 1-D (time-only) point-instrument view of one ADCP range bin.
-
-    Slices the parent dataset at *bin_idx* along ``P.ADCP_BIN_DIM`` and replaces the
-    transducer ``pressure`` with ``bin_pressure[:, bin_idx]`` so that this entry sits at
-    the correct depth in the stack.  Computes derived scalar velocity quantities::
-
-        current_speed      = sqrt(east_velocity² + north_velocity²)  [m s⁻¹, always ≥ 0]
-        current_direction  = atan2(east, north) mod 360               [degrees, 0 = N CW]
-
-    **Direction convention**: the direction *toward which* the water flows, clockwise
-    from True North.  This is the oceanographic convention (opposite to meteorological
-    "wind from").  0° = northward flow, 90° = eastward flow.
-
-    Instrument-head variables (``temperature``, ``heading``, ``pitch``, ``roll`` — see
-    ``_ADCP_HEAD_VARS``) are dropped because they belong to the transducer location, not
-    to an individual velocity bin.  Variables retaining non-time dimensions (e.g.
-    beam-dimension arrays ``amplitude``, ``correlation``, ``percent_good``) are also
-    dropped; they cannot be represented as point-instrument time series without
-    ambiguity about which beam to keep.
-
-    The result is compatible with ``_nearest_subsample`` and ``_linear_interp`` because
-    every remaining data variable has dimension ``("time",)`` or is scalar.
-
-    Parameters
-    ----------
-    ds_parent : xr.Dataset
-        Full ADCP stage-3 dataset.
-    bin_idx : int
-        Zero-based index along the bin dimension (``P.ADCP_BIN_DIM``).  Index 0 is
-        the bin nearest the transducer face.
-
-    Returns
-    -------
-    xr.Dataset
-        Single-bin time-series dataset with ``pressure`` set to ``bin_pressure``
-        at this bin and ``current_speed`` / ``current_direction`` added.
-
-    """
-    bin_dim = P.ADCP_BIN_DIM
-    ds_bin = ds_parent.isel({bin_dim: bin_idx}, drop=True)
-
-    # Drop head-only variables — these belong in the separate head entry
-    head_to_drop = [v for v in _ADCP_HEAD_VARS if v in ds_bin.data_vars]
-    if head_to_drop:
-        ds_bin = ds_bin.drop_vars(head_to_drop)
-
-    if "bin_pressure" in ds_bin.data_vars:
-        ds_bin = ds_bin.assign(
-            pressure=xr.Variable(
-                "time",
-                ds_bin["bin_pressure"].values.astype(np.float32),
-                {"units": "dbar", "long_name": "Pressure at ADCP bin"},
-            )
-        ).drop_vars("bin_pressure")
-
-    if "east_velocity" in ds_bin.data_vars and "north_velocity" in ds_bin.data_vars:
-        e = ds_bin["east_velocity"].values.astype(float)
-        n = ds_bin["north_velocity"].values.astype(float)
-        ds_bin = ds_bin.assign(
-            current_speed=xr.Variable(
-                "time",
-                np.hypot(e, n).astype(np.float32),
-                {"units": "m s-1", "long_name": "Current speed"},
-            ),
-            current_direction=xr.Variable(
-                "time",
-                (np.degrees(np.arctan2(e, n)) % 360.0).astype(np.float32),
-                {
-                    "units": "degrees",
-                    "long_name": "Current direction (oceanographic, 0=N clockwise)",
-                },
-            ),
-        )
-
-    to_drop = [v for v in ds_bin.data_vars if any(d != "time" for d in ds_bin[v].dims)]
-    if to_drop:
-        ds_bin = ds_bin.drop_vars(to_drop)
-
-    return ds_bin
-
-
-def _nearest_subsample(
-    ds: xr.Dataset,
-    common_time: np.ndarray,
-    half_window_s: float,
-) -> Dict[str, np.ndarray]:
-    """Return nearest-neighbour values on common_time within ±half_window_s seconds.
-
-    Returns a dict {varname: array} with NaN where no sample falls within the window.
-    """
-    src_t = _times_to_float(ds["time"].values)
-    tgt_t = _times_to_float(common_time)
-    n = len(common_time)
-    half_ns = half_window_s * 1e9  # ns
-
-    idx = np.searchsorted(src_t, tgt_t)
-    result: Dict[str, np.ndarray] = {}
-    for vname in STACK_VARS:
-        if vname not in ds.data_vars:
-            result[vname] = np.full(n, np.nan)
-            continue
-        src_v = ds[vname].values.astype(np.float64)
-        if vname not in _STACK_RAW:
-            src_v = _apply_qc_mask(src_v, ds, vname)
-        out = np.full(n, np.nan)
-        for i, (t_tgt, k) in enumerate(zip(tgt_t, idx)):
-            # Check candidates at k-1 and k
-            best_dt = np.inf
-            best_v = np.nan
-            for j in [k - 1, k]:
-                if 0 <= j < len(src_t):
-                    dt = abs(src_t[j] - t_tgt)
-                    if dt < best_dt and dt <= half_ns:
-                        best_dt = dt
-                        best_v = src_v[j]
-            out[i] = best_v
-        result[vname] = out
-    return result
-
-
-def _linear_interp(
-    ds: xr.Dataset,
-    common_time: np.ndarray,
-) -> Dict[str, np.ndarray]:
-    """Linearly interpolate all stack variables onto common_time; NaN outside data range."""
-    src_t = _times_to_float(ds["time"].values)
-    tgt_t = _times_to_float(common_time)
-    result: Dict[str, np.ndarray] = {}
-    for vname in STACK_VARS:
-        if vname not in ds.data_vars:
-            result[vname] = np.full(len(common_time), np.nan)
-            continue
-        src_v = ds[vname].values.astype(np.float64)
-        if vname not in _STACK_RAW:
-            src_v = _apply_qc_mask(src_v, ds, vname)
-        valid = np.isfinite(src_v)
-        if valid.sum() < 2:
-            result[vname] = np.full(len(common_time), np.nan)
-            continue
-        if vname.endswith("_qc"):
-            # QC flag arrays must not be linearly interpolated — use nearest valid
-            src_t_v = src_t[valid]
-            src_v_v = src_v[valid]
-            nn_idx = np.clip(np.searchsorted(src_t_v, tgt_t), 0, len(src_t_v) - 1)
-            result[vname] = src_v_v[nn_idx]
-        else:
-            result[vname] = np.interp(
-                tgt_t, src_t[valid], src_v[valid], left=np.nan, right=np.nan
-            )
-    return result
 
 
 class MooringStacker:
@@ -564,8 +155,6 @@ class MooringStacker:
             np.timedelta64(dt_seconds, "s"),
         ).astype("datetime64[ns]")
         n_time = len(common_time)
-
-        from .utilities import extract_inline_instruments
 
         instrument_list = list(
             mooring_config.get("clamp", mooring_config.get("instruments", []))
@@ -831,7 +420,7 @@ class MooringStacker:
             try:
                 import gsw
 
-                lat, lon = _parse_latlon(mooring_config)
+                lat, lon = parse_latlon(mooring_config)
                 T_arr = stacked["temperature"]
                 SP_arr = stacked["salinity"]
                 P_arr = stacked["pressure"]
@@ -1086,277 +675,6 @@ class MooringStacker:
         ds_out = drop_all_zero_vars(ds_out, ["amplitude_beam", "analog_input_"])
         # OceanSITES convention: time is the unlimited (first) dimension.
         ds_out = ds_out.transpose("time", "N_LEVELS")
-        ds_out = cast_output_dtypes(ds_out)
-        _enc = {
-            v: {"zlib": True, "complevel": 5}
-            for v in ds_out.data_vars
-            if ds_out[v].dtype.kind not in ("O", "U", "S")
-        }
-        ds_out.to_netcdf(output_path, encoding=_enc)
-        _status("file", self._rel(output_path))
-        return True
-
-
-class MooringGridder:
-    """Step 2: vertically interpolate stacked instruments onto a pressure grid → ``_grid.nc``."""
-
-    def __init__(
-        self,
-        base_dir: Optional[str] = None,
-        *,
-        proc_dir: Optional[str] = None,
-    ) -> None:
-        """Interpolate a mooring stack onto a regular pressure grid.
-
-        Reads ``{mooring}_stack.nc`` (produced by :class:`MooringStacker`) and
-        interpolates all scalar variables onto a uniform pressure axis defined by
-        ``p_start``, ``p_end``, and ``dp`` (in dbar).  Writes
-        ``{mooring}_grid.nc`` in the mooring proc directory.
-
-        Parameters
-        ----------
-        base_dir : str, optional
-            Legacy: cruise-level base directory containing a ``proc/`` subdirectory.
-        proc_dir : str, optional
-            Cruise-level processed output directory. Pipeline appends ``/{mooring}/``.
-
-        """
-        if base_dir is not None:
-            self.base_dir: Optional[Path] = Path(base_dir)
-            self._proc_dir: Optional[Path] = None
-            self._legacy = True
-        else:
-            self.base_dir = None
-            self._proc_dir = Path(proc_dir) if proc_dir else None
-            self._legacy = False
-
-    def _resolve_proc_dir(self, mooring_name: str) -> Path:
-        """Return the mooring-level proc directory."""
-        if not self._legacy and self._proc_dir is not None:
-            return self._proc_dir / mooring_name
-        return _get_proc_dir(self.base_dir, mooring_name)
-
-    def _rel(self, path: Path) -> str:
-        """Return a short display path relative to base_dir or proc_dir."""
-        for root in (r for r in (self.base_dir, self._proc_dir) if r):
-            try:
-                return str(path.relative_to(root))
-            except ValueError:
-                continue
-        return path.name
-
-    def grid(
-        self,
-        mooring_name: str,
-        p_start: float = 200.0,
-        p_end: float = 1000.0,
-        dp: float = 20.0,
-        force: bool = False,
-    ) -> bool:
-        """Interpolate the stacked mooring dataset onto a regular pressure grid.
-
-        Reads ``{mooring}_stack.nc``, interpolates all variables onto a pressure
-        axis from *p_start* to *p_end* in steps of *dp* dbar, and writes
-        ``{mooring}_grid.nc``.
-
-        Returns True on success, False on error.
-        """
-        proc_dir = self._resolve_proc_dir(mooring_name)
-        merge_path = proc_dir / f"{mooring_name}_stack.nc"
-        output_path = proc_dir / f"{mooring_name}_grid.nc"
-
-        try:
-            stack_found = merge_path.exists()
-            output_found = output_path.exists()
-        except (TimeoutError, OSError) as exc:
-            print("ERROR: Cannot access data drive — is it connected?")
-            print(f"       Path: {merge_path.parent}")
-            print(f"       ({type(exc).__name__}: {exc})")
-            return False
-
-        if not stack_found:
-            print(f"ERROR: Stack file not found: {merge_path}")
-            print("       Run 'oceanarray stack' first.")
-            return False
-
-        if output_found and not force:
-            _status("skip", self._rel(output_path))
-            return True
-
-        p_grid = np.arange(p_start, p_end + dp * 0.5, dp)
-        n_p = len(p_grid)
-
-        try:
-            ds = xr.open_dataset(merge_path).load()
-        except (TimeoutError, OSError) as exc:
-            print("ERROR: Cannot read stack file — is the data drive connected?")
-            print(f"       ({type(exc).__name__}: {exc})")
-            return False
-        n_time = ds.sizes["time"]
-        n_instr = ds.sizes["N_LEVELS"]
-
-        print(
-            f"Gridding {n_instr} instruments → {n_p} pressure levels "
-            f"({p_start:.0f}:{dp:.0f}:{p_end:.0f} dbar) × {n_time} time steps"
-        )
-
-        if "pressure" not in ds.data_vars:
-            print("ERROR: 'pressure' not found in stack file — cannot grid vertically")
-            ds.close()
-            return False
-
-        pressure = ds["pressure"].values.astype(np.float64)  # (time, N_LEVELS)
-        # Variables that are meaningful at the stacked per-instrument level but
-        # should not be interpolated onto the pressure grid.  Instrument-frame
-        # and beam-frame velocities are excluded because the XYZ→ENU rotation
-        # has already been applied; gridding the pre-rotation components would
-        # be misleading.  Diagnostic quantities (amplitude, correlation,
-        # battery) are per-sensor and do not have a physical meaning on a
-        # spatially-interpolated grid.
-        _GRID_EXCLUDE: frozenset = frozenset(
-            {
-                "velocity_x",
-                "velocity_y",
-                "velocity_z",
-                "velocity_beam1",
-                "velocity_beam2",
-                "velocity_beam3",
-                "amplitude",
-                "correlation",
-                "battery_voltage",
-                "velocity_flag",  # flag array, not a gridded physics variable
-                "tilt_from_pressure",  # per-instrument diagnostic, not gridded
-                "tilt_pressure_ref_hab",
-                "heading",  # instrument-frame orientation — not meaningful on a pressure grid
-                "pitch",
-                "roll",
-            }
-        )
-        grid_vars = [
-            v
-            for v in ds.data_vars
-            if v != "pressure"
-            and ds[v].dims == ("time", "N_LEVELS")
-            and not v.endswith("_qc")
-            and v not in _GRID_EXCLUDE
-        ]
-
-        stacked: Dict[str, np.ndarray] = {
-            v: np.full((n_p, n_time), np.nan) for v in grid_vars
-        }
-        var_data = {v: ds[v].values.astype(np.float64) for v in grid_vars}
-
-        # QC masking before vertical interpolation.
-        # T/S/P: NaN any non-finite values (already masked at stack time).
-        # Velocity: use velocity_flag from the stack to NaN suspect/bad samples.
-        #   Flag 3 (suspect), 4 (bad), 9 (missing) → NaN before interpolation.
-        _GRIDDER_TSQC = {"temperature", "conductivity", "salinity"}
-        for _v in grid_vars:
-            if _v in _GRIDDER_TSQC:
-                var_data[_v][~np.isfinite(var_data[_v])] = np.nan
-
-        _vel_vars = {
-            "east_velocity",
-            "north_velocity",
-            "up_velocity",
-            "current_speed",
-            "current_direction",
-        }
-        if "velocity_flag" in ds.data_vars:
-            _vflag = ds["velocity_flag"].values.astype(np.float64)
-            _bad_vel = np.isin(np.round(_vflag).astype(np.int8), [3, 4, 9])
-            for _v in _vel_vars:
-                if _v in var_data:
-                    var_data[_v][_bad_vel] = np.nan
-        else:
-            for _v in _vel_vars:
-                if _v in var_data:
-                    var_data[_v][~np.isfinite(var_data[_v])] = np.nan
-
-        # ADCP standalone QC variables — mask velocity before gridding.
-        # seabed_qc removes bins at/below the seafloor; percent_good_qc and
-        # error_velocity_qc remove poor-quality pings.  All three use flag >= 3
-        # (suspect or worse) as the NaN threshold, consistent with how the
-        # rose diagram and stack report mask these variables.
-        for _adcp_qc in ("seabed_qc", "percent_good_qc", "error_velocity_qc"):
-            if _adcp_qc not in ds.data_vars:
-                continue
-            _qc = ds[_adcp_qc].values  # (time, N_LEVELS) after stack transpose
-            _bad = np.isin(np.round(_qc).astype(np.int8), [3, 4, 9])
-            for _v in _vel_vars:
-                if _v in var_data:
-                    var_data[_v][_bad] = np.nan
-
-        for t in range(n_time):
-            p_col = pressure[t, :]
-            p_valid_mask = np.isfinite(p_col)
-            if p_valid_mask.sum() < 2:
-                continue
-
-            for vname in grid_vars:
-                v_col = var_data[vname][t, :]
-                both_valid = p_valid_mask & np.isfinite(v_col)
-                if both_valid.sum() < 2:
-                    continue
-                p_v = p_col[both_valid]
-                v_v = v_col[both_valid]
-                sort_idx = np.argsort(p_v)
-                stacked[vname][:, t] = np.interp(
-                    p_grid,
-                    p_v[sort_idx],
-                    v_v[sort_idx],
-                    left=np.nan,
-                    right=np.nan,
-                )
-
-        # Build output dataset; skip variables that are entirely NaN
-        data_vars: Dict = {}
-        for vname in grid_vars:
-            if not np.all(np.isnan(stacked[vname])):
-                a = dict(ds[vname].attrs)
-                a["vertical_interpolation"] = (
-                    f"linear in pressure; no extrapolation outside {p_start:.0f}–{p_end:.0f} dbar"
-                )
-                data_vars[vname] = xr.Variable(
-                    ("time", "pressure"), stacked[vname].T, attrs=a
-                )
-
-        ds_out = xr.Dataset(
-            data_vars,
-            coords={
-                "time": ds["time"],
-                "pressure": xr.Variable(
-                    "pressure",
-                    p_grid,
-                    {
-                        "units": "dbar",
-                        "long_name": "sea water pressure",
-                        "standard_name": "sea_water_pressure",
-                        "axis": "Z",
-                        "positive": "down",
-                    },
-                ),
-            },
-        )
-
-        prior_history = ds.attrs.get("history", "")
-        ds_out.attrs.update(
-            {
-                **{k: v for k, v in ds.attrs.items() if k not in ("history",)},
-                "p_start_dbar": p_start,
-                "p_end_dbar": p_end,
-                "dp_dbar": dp,
-                "history": (
-                    prior_history
-                    + f"; Step 2 grid: linear interpolation in pressure onto {dp:.0f} dbar pressure grid "
-                    f"({p_start:.0f}–{p_end:.0f} dbar); no extrapolation"
-                ),
-            }
-        )
-
-        ds.close()
-        if output_path.exists():
-            output_path.unlink()
         ds_out = cast_output_dtypes(ds_out)
         _enc = {
             v: {"zlib": True, "complevel": 5}
