--- oceanarray/transports.py
@@ -1,1 +1,0 @@
-"""Array-level transport computations across multiple moorings (work in progress)."""
